+++ src/app/app-routing.module.ts
@@ -1,13 +1,18 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { NotFoundComponent } from './modules/core/components';
+import { NotFoundComponent, CommentsComponent } from './modules/core/components';
 import { CoreDemoComponent } from './modules/demo/components';
 
 const routes: Routes = [
     {
         path: 'demo',
         component: CoreDemoComponent
+    },
+    {
+        path: 'comments',
+        component: CommentsComponent,
+        outlet: 'comments'
     },
     {
         path: '**',