+++ src/app/models/index.ts
@@ -4,3 +4,4 @@ export * from './product-category.enum';
 export * from './product.model';
 export * from './configuration.model';
 export * from './named-entity.model';
+export * from './comment.model';