+++ src/app/services/index.ts
@@ -3,6 +3,7 @@ export * from './product/product.service';
 export * from './local-storage/local-storage.service';
 export * from './config-options/config-options.service';
 export * from './constants/constants.service';
+export * from './comment/comment.service';
 
 export * from './generator/generator.service';
 export * from './generator/generator.factory';